nano-research.csl
The file has been created using a template from "ieee.csl". As to the citation of Journal articles and conference proceedings are concerned, it perfectly matches with the requirement of the journal (Nano Research). However, Books, Chapters, thesis, and Patents need  to be added by the contributors. I hope whoever uses this one will benifit from it.
Thanking you in anticipation!
Sincerely
Fasihullah Khan
Dongguk University, Republic of Korea.